requirements.txt
python-dotenv>=1.0
pydantic>=2.5
sqlalchemy>=2.0
psycopg2-binary>=2.9
neo4j>=5.14
pandas>=2.1
numpy>=1.26
pdfplumber>=0.11
beautifulsoup4>=4.12
lxml>=4.9
langchain>=0.2
openai>=1.0
tiktoken>=0.7
qdrant-client>=1.8
pgvector>=0.2
scikit-learn>=1.3
requests>=2.31
pytest>=7.4
black>=23.12
torch>=2.0
sentence-transformers>=2.3
jpype1>=1.4.0
